Fix Initilializing Multiple configuration Contexts when load testing API Gateway With WSClient Enable
