+++ components/apimgt/org.wso2.carbon.apimgt.gateway/src/main/java/org/wso2/carbon/apimgt/gateway/internal/APIHandlerServiceComponent.java
@@ -16,6 +16,9 @@
 
 package org.wso2.carbon.apimgt.gateway.internal;
 
+import org.apache.axis2.AxisFault;
+import org.apache.axis2.context.ConfigurationContext;
+import org.apache.axis2.context.ConfigurationContextFactory;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.osgi.framework.BundleContext;
@@ -32,6 +35,7 @@ import org.wso2.carbon.apimgt.gateway.throttling.util.KeyTemplateRetriever;
 import org.wso2.carbon.apimgt.impl.APIConstants;
 import org.wso2.carbon.apimgt.impl.APIManagerConfiguration;
 import org.wso2.carbon.apimgt.impl.APIManagerConfigurationService;
+import org.wso2.carbon.base.ServerConfiguration;
 import org.wso2.carbon.utils.Axis2ConfigurationContextObserver;
 import org.wso2.carbon.utils.CarbonUtils;
 import org.wso2.carbon.utils.ConfigurationContextService;
@@ -57,23 +61,27 @@ public class APIHandlerServiceComponent {
     private ServiceRegistration registration;
 
     protected void activate(ComponentContext context) {
-    	BundleContext bundleContext = context.getBundleContext();
+        BundleContext bundleContext = context.getBundleContext();
         if (log.isDebugEnabled()) {
             log.debug("API handlers component activated");
         }
-        clientPool = APIKeyValidatorClientPool.getInstance();
-        thriftClientPool = ThriftKeyValidatorClientPool.getInstance();
+        try {
+            ConfigurationContext ctx = ConfigurationContextFactory.createConfigurationContextFromFileSystem
+                    (getClientRepoLocation(), getAxis2ClientXmlLocation());
+            ServiceReferenceHolder.getInstance().setAxis2ConfigurationContext(ctx);
 
-        String filePath = getFilePath();
-		try {
-			configuration.load(filePath);
+            clientPool = APIKeyValidatorClientPool.getInstance();
+            thriftClientPool = ThriftKeyValidatorClientPool.getInstance();
 
-			String gatewayType = configuration.getFirstProperty(APIConstants.API_GATEWAY_TYPE);
-			if ("Synapse".equalsIgnoreCase(gatewayType)) {
-			  //Register Tenant service creator to deploy tenant specific common synapse configurations
-			  TenantServiceCreator listener = new TenantServiceCreator();
-			  bundleContext.registerService(
-			          Axis2ConfigurationContextObserver.class.getName(), listener, null);
+            String filePath = getFilePath();
+            configuration.load(filePath);
+
+            String gatewayType = configuration.getFirstProperty(APIConstants.API_GATEWAY_TYPE);
+            if ("Synapse".equalsIgnoreCase(gatewayType)) {
+                //Register Tenant service creator to deploy tenant specific common synapse configurations
+                TenantServiceCreator listener = new TenantServiceCreator();
+                bundleContext.registerService(
+                        Axis2ConfigurationContextObserver.class.getName(), listener, null);
 
                 if (configuration.getThrottleProperties().isEnabled()) {
                     ThrottleDataHolder throttleDataHolder = new ThrottleDataHolder();
@@ -97,7 +105,7 @@ public class APIHandlerServiceComponent {
                     //Following method will initialize JMS listnet and listen all updates and keep throttle data map
                     // up to date
                     //start web service throttle data retriever as separate thread and start it.
-                    if (configuration.getThrottleProperties().getBlockCondition().isEnabled()){
+                    if (configuration.getThrottleProperties().getBlockCondition().isEnabled()) {
                         BlockingConditionRetriever webServiceThrottleDataRetriever = new
                                 BlockingConditionRetriever();
                         webServiceThrottleDataRetriever.startWebServiceThrottleDataRetriever();
@@ -107,9 +115,9 @@ public class APIHandlerServiceComponent {
                     }
                 }
             }
-		} catch (APIManagementException e) {
-			log.error("Error while initializing the API Gateway (APIHandlerServiceComponent) component", e);
-		}
+        } catch (APIManagementException | AxisFault e) {
+            log.error("Error while initializing the API Gateway (APIHandlerServiceComponent) component", e);
+        }
     }
 
     protected void deactivate(ComponentContext context) {
@@ -159,5 +167,14 @@ public class APIHandlerServiceComponent {
     protected void setConfiguration(APIManagerConfiguration configuration) {
         this.configuration = configuration;
     }
-
+    protected String getAxis2ClientXmlLocation() {
+        String axis2ClientXml = ServerConfiguration.getInstance().getFirstProperty("Axis2Config" +
+                ".clientAxis2XmlLocation");
+        return axis2ClientXml;
+    }
+    protected String getClientRepoLocation() {
+        String axis2ClientXml = ServerConfiguration.getInstance().getFirstProperty("Axis2Config" +
+                ".ClientRepositoryLocation");
+        return axis2ClientXml;
+    }
 }
