+++ components/apimgt/org.wso2.carbon.apimgt.gateway/src/test/java/org/wso2/carbon/apimgt/gateway/handlers/security/keys/APIKeyValidatorClientWrapper.java
@@ -8,13 +8,4 @@ public class APIKeyValidatorClientWrapper extends APIKeyValidatorClient {
     public APIKeyValidatorClientWrapper() throws APISecurityException {
     }
 
-    @Override
-    protected String getAxis2ClientXmlLocation() {
-        return "src/test/resources/axis2_client.xml";
-    }
-
-    @Override
-    protected String getClientRepoLocation() {
-        return "src/test/resources/client";
-    }
 }
